requirements-3.11.txt
albumentations==1.3.1
asttokens==2.4.1
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
exceptiongroup==1.1.3
executing==2.0.1
fonttools==4.44.0
imageio==2.32.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.1
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.2
parso==0.8.3
Pillow==10.0.1
platformdirs==4.0.0
prompt-toolkit==3.0.40
psutil==5.9.6
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
qudida==0.0.4
rich==13.6.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.2.0
tifffile==2023.9.26
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.9
